[Prototype] Added first implementation of player movement
Right now it's not working at all, but I'll still push it to comment it.
I think I found a bug, reachable nodes should be just the ones around the current one, but for some reason they are 300+. Need to check it.
Probably the movement is broken (also) because of this.